@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.pavelsikun.seekbarpreference.sample"
-        minSdkVersion 7
-        targetSdkVersion 23
+        minSdkVersion 21
+        targetSdkVersion 29
         versionCode 6
         versionName "6.0"
     }
@@ -20,11 +19,11 @@ android {
 }
 
 dependencies {
-    compile project(':library')
+    implementation project(':library')
     //    for your projects use:
 //        compile 'com.pavelsikun:material-seekbar-preference:2.3.0'
 
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:cardview-v7:23.4.0'
-    compile 'com.android.support:preference-v7:23.4.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.preference:preference:1.1.1'
 }
